r1620: Began adding commands to make unix_file_bd not use the host OS's cacheing to help give us a better feel for what our current performance is when we talk directly to physical BDs.
open() is now passed the O_DSYNC flag to make data synchronous; however,
linux implements O_DSYNC with O_SYNC (e.g. also sync access time) and I do
not know if MacOSX has O_DSYNC or O_SYNC.

Added a TODO for doing direct io (to skip the host OS's caches) with
possible ways to do it for linux and osx that appear to be platform dependent.

Added a NOTE for unix_file_bd_sync() that OSX has a fcntl() command to flush
a drive's buffer, addressing this function's current #warnings on this issue.

Thoughts on and improvements to these changes will be most welcome!