r1685: Buffer overflow fix.
josfs_get_dirent(buf) could write filename data past the end of the provided
buffer because it could copy more than it calculated to be safe. Fixed
by correcting the copy length from the maximal to the actual length;
the actual length was and is used for the amount-safe calculation.

With the buffer on the stack, this function overwrote other stack
variables. It wrote zeros.

This bug is not in ufs_base.c.