r2243: Updated revision_slice_chdesc_is_ready() to recurse using the heap instead of the stack. This function was using more than 8kB of stack; with this reivision I can extract linux-2.6.15 to uukfsd using a 6kB stack.
This function is by far our most time consuming function at the present.
malloc() slowed this function, and thus kfsd, down a good bit so this change
uses a static buffer large enough for all depths I've seen and it falls
back to a size-doubling malloc() if this static buffer is exceeded.
This does help, but it looks like we still take a decent speed hit for this
change.