r2696: Allow byte chdescs to span an entire block instead of only an atomic unit. Restricting chdescs to atomic units creates additional chdescs and chmetadescs and is this restriction is unnecessary.
As of r2694 the #2 memory user are chdesc_ts. This revision allows the
creation of a large byte block change to create a single chdesc for up to an
entire block instead of one for each atomic unit; reducing the number of
change descriptors for a UFS block from 4 to 1 or for an ext2 block from 8 to 1.
This also reduces the number of created chmetadesc_ts.

For a 16384 block write back cache with UFS, this reduces chdesc_t memory use
from 35MB to 25MB and chmetadesc_t memory use from 20MB to 10-15MB.

Because this change simplified byte chdesc creation, it also allowed byte
chdescs with omitted data to be applied directly to the ddesc, avoiding
the memdup, memxchg, and immediate free that was previously needed.