r3506: Change directory creation code so that a new directory data block (and any following dirents) can be written in one disk write instead of two. This patch has been waiting for checkin for sometime now; the recent ext2 allocation enhancements (r3499 and r3500) make the disk requests generated to satisfy these updated dependencies feasible.
Though this checkin reduces the number of blocks written multiple times
and it will allow a following checkin to further increase patch merging and
reduce the number of soft patches, it does increase the number of disk requests
needed to write a tree of directories, from 2 to the depth of the tree.
We may want to add dynamic dependency optimization to resolve this tradeoff.