r3633: Getting rid of repeated calls to get_write_head() and replacing them with pointer derefecences.
Now get_write_head() returns a chdesc_t **, which should be dereferenced to get
the previous return value of get_write_head() whenever necessary later. This
means that modules which do nontrivial things with the write head (i.e. the
journal module) now must deal with the fact that they are no longer explicitly
asked for the value whenever another module wants it. For the journal module,
this means that it must always have a current transaction instead of creating
one on demand as it did before.