Fix kernel recursive inline request compile error by removing the request.
The recursive inline request for patch_may_have_before() is not handled
by gcc 4.2. Apparently we used to think that gcc:
- <4 ignored the inline request (so left it in)
- 4.0 and 4.1 could not handle the inline request (so omitted it)
- >=4.2 handled the inline request (so let it)
However, it seems that this last point was incorrect.
So this commit removes the request for all compilers, since
the request was probably never used by our compilers,
Kernel Featherstitch now compiles with the Ubuntu 8.04 build tools.